functions/requirements.txt
adal==0.5.1
aiofiles==0.3.2
aiohttp==3.0.7
asn1crypto==0.24.0
astroid==1.6.2
async-timeout==2.0.0
attrs==17.4.0
azure==3.0.0
azure-batch==4.1.1
azure-common==1.1.9
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.2
azure-datalake-store==0.0.19
azure-eventgrid==0.1.0
azure-graphrbac==0.40.0
azure-keyvault==0.3.7
azure-mgmt==2.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==5.0.0
azure-mgmt-batchai==0.2.0
azure-mgmt-billing==0.1.0
azure-mgmt-cdn==2.0.0
azure-mgmt-cognitiveservices==2.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==3.0.1
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==0.3.1
azure-mgmt-containerregistry==1.0.1
azure-mgmt-containerservice==3.0.1
azure-mgmt-cosmosdb==0.3.1
azure-mgmt-datafactory==0.4.0
azure-mgmt-datalake-analytics==0.3.0
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.3.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==1.2.0
azure-mgmt-eventgrid==0.4.0
azure-mgmt-eventhub==1.2.0
azure-mgmt-hanaonazure==0.1.0
azure-mgmt-iothub==0.4.0
azure-mgmt-iothubprovisioningservices==0.1.0
azure-mgmt-keyvault==0.40.0
azure-mgmt-loganalytics==0.1.0
azure-mgmt-logic==2.1.0
azure-mgmt-machinelearningcompute==0.4.0
azure-mgmt-managementpartner==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==0.2.0
azure-mgmt-monitor==0.4.0
azure-mgmt-msi==0.1.0
azure-mgmt-network==1.7.1
azure-mgmt-notificationhubs==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-powerbiembedded==1.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-recoveryservices==0.2.0
azure-mgmt-recoveryservicesbackup==0.1.1
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.1.0
azure-mgmt-resource==1.2.2
azure-mgmt-scheduler==1.1.3
azure-mgmt-search==1.0.0
azure-mgmt-servermanager==1.2.0
azure-mgmt-servicebus==0.4.0
azure-mgmt-servicefabric==0.1.0
azure-mgmt-sql==0.8.6
azure-mgmt-storage==1.5.0
azure-mgmt-subscription==0.1.0
azure-mgmt-trafficmanager==0.40.0
azure-mgmt-web==0.34.1
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==6.1.2.9
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.1.0
azure-storage-common==1.1.0
azure-storage-file==1.1.0
azure-storage-nspkg==3.0.0
azure-storage-queue==1.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
cssselect==1.0.3
defusedxml==0.5.0
entrypoints==0.2.3
fake-useragent==0.1.10
Flask==0.12.2
Flask-Login==0.4.1
futures==3.1.1
giturlparse.py==0.0.5
grab==0.6.38
idna==2.6
idna-ssl==1.0.1
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
keyring==12.0.0
lazy-object-proxy==1.3.1
lxml==4.2.1
MarkupSafe==1.0
mccabe==0.6.1
msrest==0.4.27
msrestazure==0.4.25
multidict==4.1.0
oauthlib==2.0.7
parse==1.8.2
Pillow==5.1.0
praw==5.4.0
prawcore==0.14.0
pycodestyle==2.3.1
pycparser==2.18
pycurl==7.43.0.1
pyee==5.0.0
PyJWT==1.6.1
pylint==1.8.3
pyppeteer==0.0.17
pyquery==1.4.0
python-dateutil==2.7.2
pytils==0.3
requests==2.18.4
requests-html==0.9.0
requests-oauthlib==0.8.0
rope==0.10.7
SecretStorage==2.3.1
selection==0.0.13
six==1.11.0
tblib==1.3.2
tornado==5.0
u-msgpack-python==2.4.1
update-checker==0.16
urllib3==1.22
user-agent==0.1.9
w3lib==1.19.0
weblib==0.1.25
websockets==4.0.1
Werkzeug==0.14.1
wrapt==1.10.11
WsgiDAV==2.2.4
yapf==0.21.0
yarl==1.1.1
